specify AR tool to use
A previous commit stopped specifying the AR tool that gets used to
bootstrap make, because macOS would specify libtool which is not
compatible with make's configure script. Not specifying the tool
breaks Linux builds though. As a (hopefully temporary) hack we replace
libtool on macOS with llvm-ar which we know exists in our toolchain.